chore: split index commands
